refactor: restructure components and enhance typing
- Moved Wordlist and WordOfTheDay components to Organisms directory
- Created type definitions for Wordlist and WordOfTheDay components
- Updated Main component to use new paths for Wordlist and WordOfTheDay
- Improved typing for BlockElement and Typography components
