+++ src/Components/UI/Header/Header.tsx
@@ -3,8 +3,8 @@ import { AppContentData } from '../../../Data/AppContent';
 import { Typography } from '../../Typography/Typography';
 import { dataDays } from '../../../Data/Calendar/Days';
 import { dataMonths } from '../../../Data/Calendar/Months';
-import { IHeaderProps } from './Header.d';
 import { BlockElement } from '../BlockLevel/BlockElement';
+import { IHeaderProps } from './Header.d';
 
 const { mainHeading, date } = AppContentData.globalPageContent;
